requirements.txt
appdirs==1.4.4
certifi @ file:///work/ci_py311/certifi_1676822443922/work/certifi
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.10.7
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lit==16.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.0
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.22.1
psutil==5.9.4
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.18.0
setproctitle==1.3.2
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.14.0
